+++ app/langs/gl.php
@@ -22,12 +22,15 @@ return [
 
     'menu' => [
         'home' => 'Páxina principal',
-        'about' => 'Sobre nós',
-        'cookies' => 'Ley de cookies',
+        'about' => 'Sobre nós'
     ],
 
     'about' => 'Sobre nós',
 
     'cookies' => 'Ley de cookies',
+
+    'cookies_msn' => 'En cumprimento da Lei 34/2002, de Servizos da Sociedade da Información, lembrámoslle que este sitio web utiliza cookies co fin de mellorar os nosos servizos.',
+    'cookies_accept' => 'Aceptar',
+    'cookies_more_info' => 'Máis información',
 
 ];