+++ public/static/js/site.js
@@ -1,3 +1,16 @@
 $(function(){
-
+	$('#cookies-accept-btn').click(function(event){
+		event.preventDefault();
+		App.setCookie();
+	})
 });
+
+var App = {};
+App.cookies = {};
+
+App.setCookie  = function(){
+   myDate = new Date();
+   myDate.setTime(myDate.getTime()+(App.cookies.days*24*60*60*1000));
+   document.cookie = ''+App.cookies.name+'=True; expires=' + myDate.toGMTString()+';path=/';
+   $('.cookie-advise').fadeOut();
+}